+++ includes/class-spirit-dashboard.php
@@ -14,9 +14,8 @@
  * @subpackage Spirit_Dashboard/includes
  * @author     Vaggelis Pallis <info.vpsnak@gmail.com>
  */
-class Spirit_Dashboard
-{
-
+class Spirit_Dashboard {
+    
     /**
      * The loader that's responsible for maintaining and registering all hooks that power
      * the plugin.
@@ -26,7 +25,7 @@ class Spirit_Dashboard
      * @var      Spirit_Dashboard_Loader $loader Maintains and registers all hooks for the plugin.
      */
     protected $loader;
-
+    
     /**
      * The unique identifier of this plugin.
      *
@@ -35,7 +34,7 @@ class Spirit_Dashboard
      * @var      string $plugin_name The string used to uniquely identify this plugin.
      */
     protected $plugin_name;
-
+    
     /**
      * The current version of the plugin.
      *
@@ -44,7 +43,7 @@ class Spirit_Dashboard
      * @var      string $version The current version of the plugin.
      */
     protected $version;
-
+    
     /**
      * Define the core functionality of the plugin.
      *
@@ -54,19 +53,18 @@ class Spirit_Dashboard
      *
      * @since    0.0.1
      */
-    public function __construct()
-    {
+    public function __construct () {
         if (defined('SPIRIT_DASHBOARD_VERSION')) {
             $this->version = SPIRIT_DASHBOARD_VERSION;
         } else {
             $this->version = '0.0.1';
         }
         $this->plugin_name = 'spirit-dashboard';
-
+        
         $this->load_dependencies();
-
+        
     }
-
+    
     /**
      * Load the required dependencies for this plugin.
      *
@@ -83,31 +81,32 @@ class Spirit_Dashboard
      * @since    0.0.1
      * @access   private
      */
-    private function load_dependencies()
-    {
-
+    private function load_dependencies () {
+        
         /**
          * The class responsible for orchestrating the actions and filters of the
          * core plugin.
          */
         require_once plugin_dir_path(dirname(__FILE__)) . 'includes/class-spirit-dashboard-loader.php';
-
+        
         require_once plugin_dir_path(dirname(__FILE__)) . 'includes/rest-spirit-dashboard.php';
-
+        
+        require_once plugin_dir_path(dirname(__FILE__)) . 'includes/application/class-spirit-passwords.php';
+        
         $this->loader = new Spirit_Dashboard_Loader();
-
+        
     }
-
+    
     /**
      * Run the loader to execute all of the hooks with WordPress.
      *
      * @since    0.0.1
      */
-    public function run()
-    {
+    public function run () {
         $this->loader->run();
+        Spirit_Passwords::add_hooks();
     }
-
+    
     /**
      * The name of the plugin used to uniquely identify it within the context of
      * WordPress and to define internationalization functionality.
@@ -115,31 +114,28 @@ class Spirit_Dashboard
      * @since     0.0.1
      * @return    string    The name of the plugin.
      */
-    public function get_plugin_name()
-    {
+    public function get_plugin_name () {
         return $this->plugin_name;
     }
-
+    
     /**
      * The reference to the class that orchestrates the hooks with the plugin.
      *
      * @since     0.0.1
      * @return    Spirit_Dashboard_Loader    Orchestrates the hooks of the plugin.
      */
-    public function get_loader()
-    {
+    public function get_loader () {
         return $this->loader;
     }
-
+    
     /**
      * Retrieve the version number of the plugin.
      *
      * @since     0.0.1
      * @return    string    The version number of the plugin.
      */
-    public function get_version()
-    {
+    public function get_version () {
         return $this->version;
     }
-
+    
 }