+++ spirit-dashboard.php
@@ -16,7 +16,7 @@
  * Plugin Name:       Spirit Dashboard
  * Plugin URI:        https://github.com/vpsnak/spirit-dashboard-plugin
  * Description:       This plugin is used to add endpoints for listing and updating Wordpress core, plugins, themes, translations.
- * Version:           1.1.0
+ * Version:           1.1.1
  * Author:            Vaggelis Pallis
  * Author URI:        https://github.com/vpsnak
  * License:           GPL-2.0+
@@ -37,14 +37,15 @@ if (!defined('WPINC')) {
  * Start at version 0.0.1 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define('SPIRIT_DASHBOARD_VERSION', '1.1.0');
+define('SPIRIT_DASHBOARD_VERSION', '1.1.1');
+
+define('SPIRIT_DASHBOARD_BASE_URL', plugin_dir_url(__FILE__));
 
 /**
  * The code that runs during plugin activation.
  * This action is documented in includes/class-spirit-dashboard-activator.php
  */
-function activate_spirit_dashboard()
-{
+function activate_spirit_dashboard () {
     require_once plugin_dir_path(__FILE__) . 'includes/class-spirit-dashboard-activator.php';
     Spirit_Dashboard_Activator::activate();
 }
@@ -53,8 +54,7 @@ function activate_spirit_dashboard()
  * The code that runs during plugin deactivation.
  * This action is documented in includes/class-spirit-dashboard-deactivator.php
  */
-function deactivate_spirit_dashboard()
-{
+function deactivate_spirit_dashboard () {
     require_once plugin_dir_path(__FILE__) . 'includes/class-spirit-dashboard-deactivator.php';
     Spirit_Dashboard_Deactivator::deactivate();
 }
@@ -77,31 +77,14 @@ require plugin_dir_path(__FILE__) . 'includes/class-spirit-dashboard.php';
  *
  * @since    0.0.1
  */
-function run_spirit_dashboard()
-{
-
+function run_spirit_dashboard () {
+    
     $plugin = new Spirit_Dashboard();
     $plugin->run();
-
+    
     require_once plugin_dir_path(__FILE__) . 'includes/plugin-update-checker/plugin-update-checker.php';
-    $myUpdateChecker = Puc_v4_Factory::buildUpdateChecker(
-        'https://github.com/vpsnak/spirit-dashboard-plugin/',
-        __FILE__,
-        'spirit-dashboard'
-    );
-//
-//        var_dump(plugin_dir_path(__FILE__));
-//        var_dump(plugin_dir_path(dirname(__FILE__)));
-//        var_dump(plugin_dir_path( __DIR__ ));
-
-    //Optional: If you're using a private repository, specify the access token like this:
-//    $myUpdateChecker->setAuthentication('');
-
-    $myUpdateChecker->getVcsApi()->enableReleaseAssets();
-
-    //Optional: Set the branch that contains the stable release.
-//        $myUpdateChecker->setBranch('master');
-
+    $UpdateChecker = Puc_v4_Factory::buildUpdateChecker('https://github.com/vpsnak/spirit-dashboard-plugin/', __FILE__, 'spirit-dashboard');
+    $UpdateChecker->getVcsApi()->enableReleaseAssets();
 }
 
 run_spirit_dashboard();